image resize and rotate

+++ src/main/java/net/perspective/draw/event/behaviours/FigureItemBehaviour.java
@@ -191,9 +191,9 @@ public class FigureItemBehaviour implements ItemBehaviours {
                         context.setQuad(quad);
                     }
                 }
-                double[] flip = R2.flip(context.getQuad());
-                double cos_t = flip[0];
-                double sin_t = flip[1];
+                int[] flip = R2.flip(context.getQuad());
+                int cos_t = flip[0];
+                int sin_t = flip[1];
                 @SuppressWarnings("deprecation")
                 double t = item.getAngle() + (item.isVertical() ? -Math.PI / 2 : 0);
                 double delta = V2.norm_angle(4 * t + 2 * Math.PI) / 2;
